src/phantom/requirements.txt
tensorflow==2.12.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
python-dotenv==1.0.0
transformers==4.30.2
torch==2.0.1